+++ src/Exception/LogicExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Solarium\Exception;
+
+/**
+ * Marker Interface for Solarium logic exceptions.
+ */
+interface LogicExceptionInterface extends ExceptionInterface
+{
+}