+++ src/Exception/RuntimeExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Solarium\Exception;
+
+/**
+ * Marker Interface for Solarium runtime exceptions.
+ */
+interface RuntimeExceptionInterface extends ExceptionInterface
+{
+}