+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to the solarium library will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
+
+## [6.0.0]
+### Added
+- \Solarium\Component\Result\Facet\Buckets::getNumBuckets()
+
+### Changed
+- Thrown exceptions always implement Solarium\Exception\ExceptionInterface
+
 
 ## [6.0.0-rc.1]
 ### Added